requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiosqlite==0.20.0
python-telegram-bot==21.6
dateparser==1.2.0
apscheduler==3.10.4
pydantic-settings==2.5.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytz==2024.1
